requirements.txt
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
fastapi-cache2==0.2.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pendulum==3.1.0
pluggy==1.6.0
prometheus_client==0.22.1
pydantic==1.10.12
pydantic_core
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
databases[sqlite]
fastapi-jwt-auth-compat
PyJWT>=2.6.0
sqlalchemy